Refactor WebSocket logging and replace routes with Public endpoint; add user authentication and profile image handling
